Main Menu button hover visuals
The buttons on the main menu now grow and change to a brighter color when you hover over them with the mouse cursor. The also then return back to its normal look when the mouse leaves the button.

I also fixed an issue with the settings menu dealing with the canvas where it wasn't scaling properly.